Level Generation and navigation
Level Generation scripts setup for this project.

Fixed some issues with the Player platform not being updated.

BlockDropThrough now does nothing, keeping it as a setup note for now.

Updated StageClear/RoomClear and EnemyStageManager scripts that manage door toggles.

DoorManager now gets all doors connected to the Origin's room.

Added script to move camera to the current room the player is in.