Boss and Platform fixes
Fixed Boss floating in air if killed during float. Rigidbody and collider physics were being disabled on Death, added a short delay to allow the Boss to fall to the ground first.

Fixed Boss not Sleeping on spawn.

Updated Sprite layers for Walls, Platforms and Doors.
Walls and Doors are now on a new layer in front of the Ground layer.

Fixed collisions issues with some Room Platforms and Doors.