Added Augment select/purchase
When clicking on an Augment, it is moved to the Player AugmentInventory.

Displayed price will now show red if the Player can't afford it.

Augments can only be selected/purchased if the Player has enough money.

Augments are greyed out after purchase with a "Purchased" overlay.

Purchased Augments are moved to the "ownedAugments" pool, and will no longer be added to Shops if refreshed.