Augment Select update for duplicates
Augment overlay will now display "Purchased" once purchased and "Max Level" if the Augment is already owned, and is at max level.
If the augment is not at max level, the Level will display "Lv ??" and the level will be randomized.
"Owned" text will also be displayed to indicate the Augment is a duplicate.

Added a MaxLevel to Augments, default value is 5. Higher tiers have lower max levels.