Knockback updates and fixes
Moved Indicator from a pool to the Enemy object directly, this fixes detachment issues.

Fixed Knockback, enemy GetKnockback() was reducing the knockback by the enemy’s knockback resist and the Shielder was reducing the knockback to a negative value.
Added a check to just prevent Knockback if it reached 0 after applying the resist.

Fixed Knockback isAttacking and Flip() toggles. canMove and ToggleFlip are only reset in KnockbackReset if isAttacking is true, otherwise AttackingCO will reset these.

Fixed issues with Enemies detecting the Shop and Spawner colliders.