Minimap updates
Updated HitStop to slow down time for a short duration instead of completely stopping time. There were issues with the WaitForSecondsRealtime interfering with the PauseMenu.

Updated minimap to only show rooms if the Player has entered the room, then highlights the room if cleared.

Doors/Walls are also disabled at Start() unless it is the Starting Room. Doors/Walls and Room icons are revealed when the Player enters the rooms.