New Caustic Augment added, proc chance augments
Added Proc chance to augment descriptions. Proc chance in the base description is represented with a '$' symbol which is replaced in AugmentScript.

Added Caustic augment that applies a slow poison DoT to enemies. This DoT can stack, but has a 25% chance to proc and deals damage every 2s over 4s.