Settings update, Loading screen and audio
Fixed issue with Video settings not being displayed correctly across scenes.

Audio now fades Out when loading a new scene, and fade In when the scene is loaded.
AsyncSceneLoader script now unloads Player and Level scene when returning to Menu.

Loading screen overlay added. Screen is displayed as long as the scene is loading.