Menu pausing, Loading fixes
Music is now paused when the game is.
PlayScheduled timer is also paused.

Added delay when starting a loading screen to allow audio to fully fade out.

Main Menu input disabled once Play is selected.
Input is also disabled when the Menu is being loaded from the in-game Pause Menu.